Fix: Add explicit cast for DesignWidth in thumbnail creation
In `MainPage/MainWindow.xaml.cs`, within `LoadChapterListAsync`, I added an explicit `(int)` cast when assigning `ChapterListElement.DesignWidth` (which is a `double`) to `BitmapImage.DecodePixelWidth` (which expects an `int`).

This corrects a type mismatch introduced when `ChapterListElement.DesignWidth` was changed to `double` to resolve a XAML `InvalidCastException` with `WrapPanel.ItemWidth`.

+++ MainPage/MainWindow.xaml.cs
@@ -71,7 +71,7 @@ namespace Reader
                         BitmapImage thumbnail = new BitmapImage();
                         thumbnail.BeginInit();
                         thumbnail.UriSource = imageSourceUri;
-                        if (width > height) { thumbnail.DecodePixelWidth = ChapterListElement.DesignWidth; }
+                        if (width > height) { thumbnail.DecodePixelWidth = (int)ChapterListElement.DesignWidth; }
                         else { thumbnail.DecodePixelHeight = ChapterListElement.ImageHeight; }
                         thumbnail.CreateOptions = BitmapCreateOptions.None;
                         thumbnail.CacheOption = BitmapCacheOption.OnLoad;
